+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/Option.java
@@ -228,6 +228,12 @@ public enum Option {
      */
     INLINE_ALL_SCHEMAS(InlineSchemaModule::new, null, Option.DEFINITIONS_FOR_ALL_OBJECTS, Option.DEFINITION_FOR_MAIN_SCHEMA),
     /**
+     * Generally, keys in the collected "definitions"/"$defs" are ensured to be URI compatible but may include parentheses and commas for listing type
+     * parameters. By enabling this option, these parentheses and commas will be removed to conform with a reduced set of characters, e.g. as expected
+     * by OpenAPI: following the regular expression {@code ^[a-zA-Z0-9\.\-_]+$}.
+     */
+    PLAIN_DEFINITION_KEYS(null, null),
+    /**
      * Whether as the last step of the schema generation, unnecessary "allOf" elements (i.e. where there are no conflicts/overlaps between the
      * contained sub-schemas) should be merged into one, in order to make the generated schema more readable. This also applies to manually added
      * "allOf" elements, e.g. through custom definitions or attribute overrides.